requirements.txt
altair
pandas
streamlit
diffusers 
transformers 
omegaconf 
safetensors 
accelerate 
pillow
